Update acronymizer to use first word for unique pilots
For cards with long names which are unique I suggest we try to use the first word if possible.

We’ll get `Biggs`, `Wes` and `Corran` instead of `BD`, `WJ` and `CH`. It doesn’t always work perfectly (`Darth` instead of `Vader`, `Pure` instead of `Sabbac`) but if we’d like to fix that we’d need to maintain a list of acronyms which might be a bit too much.

I also removed quotes from names as it takes up space on the base/dial without really adding anything. So now `”Whisper”` will be shown as `Whisper` instead of `W` (as it’s no longer more than 8 characters)